backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
openpyxl>=3.1.0
# The model was trained with scikit-learn 1.5.1. Ideally use the same version,
# but 1.5.1 has no wheels for Python 3.14. A compatibility shim in
# pump_api/services.py patches the missing _RemainderColsList class so that
# newer sklearn (1.8+) can still load the old pickle.
scikit-learn>=1.5.1
xgboost>=2.0.0
